backend/requirements.txt
fastapi
fastapi-sqlalchemy
sqlalchemy[full]
uvicorn[standard]
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
py-healthcheck
pydantic
databases
asyncpg